% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeInSQL.R
\name{connectionInfo}
\alias{connectionInfo}
\title{Execute a function in SQL}
\usage{
connectionInfo(driver = "SQL Server", server = "localhost",
  database = "master", uid = NULL, pwd = NULL)
}
\arguments{
\item{driver}{The driver to use for the connection - defaults to SQL Server}

\item{server}{The server to connect to - defaults to localhost}

\item{database}{The database to connect to - defaults to master}

\item{uid}{The user id for the connection. If uid is NULL, default to Trusted Connection}

\item{pwd}{The password for the connection. If uid is not NULL, pwd is required}
}
\value{
A fully formed connection string
}
\description{
Execute a function in SQL
}
\examples{
\dontrun{

connectionInfo()
[1] "Driver={SQL Server};Server=localhost;Database=master;Trusted_Connection=Yes;"

connectionInfo(server="ServerName", database="AirlineTestDB", uid="username", pwd="pass")
[1] "Driver={SQL Server};Server=ServerName;Database=AirlineTestDB;uid=username;pwd=pass;"
}


}
