% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeInSQL.R
\name{executeFunctionInSQL}
\alias{executeFunctionInSQL}
\title{Execute a function in SQL}
\usage{
executeFunctionInSQL(connectionString, func, ..., inputDataQuery = "",
  getScript = FALSE)
}
\arguments{
\item{connectionString}{character string. The connectionString to the database}

\item{func}{closure. The function to execute}

\item{...}{A named list of arguments to pass into the function}

\item{inputDataQuery}{character string. A string to query the database.
The result of the query will be put into a data frame into the first argument in the function}

\item{getScript}{boolean. Return the tsql script that would be run on the server instead of running it}
}
\value{
The returned value from the function
}
\description{
Execute a function in SQL
}
\examples{
\dontrun{
connection <- connectionInfo(database = "AirlineTestDB")

foo <- function(in_df, arg) {
    list(data = in_df, value = arg)
}
executeFunctionInSQL(connection, foo, arg = 12345,
                     inputDataQuery = "SELECT top 1 * from airline5000")
}


}
\seealso{
\code{\link{executeScriptInSQL}} to execute a script file instead of a function in SQL
}
