% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedProcedure.R
\name{executeSproc}
\alias{executeSproc}
\title{Execute a Stored Procedure}
\usage{
executeSproc(connectionString, name, ..., getScript = FALSE)
}
\arguments{
\item{connectionString}{character string. The connectionString for the database with the stored procedure}

\item{name}{character string. The name of the stored procedure in the database to execute}

\item{...}{named list. Parameters to pass into the procedure. These MUST be named the same as the arguments to the function.}

\item{getScript}{boolean. Return the tsql script that would be run on the server instead of running it}
}
\description{
Execute a Stored Procedure
}
\section{Warning}{

Even though you can create stored procedures with output parameters, you CANNOT execute them
using this utility due to limitations of RODBC.
}

\examples{
\dontrun{
connectionString <- connectionInfo()

dropSproc(connectionString, "fun")

func <- function(arg1) {return(data.frame(hello = arg1))}
createSprocFromFunction(connectionString, name = "fun",
                        func = func, inputParams = list(arg1="character"))

if (checkSproc(connectionString, "fun")) {
    print("Function 'fun' exists!")
    executeSproc(connectionString, "fun", arg1="WORLD")
}
}
}
\seealso{
{
\code{\link{createSprocFromFunction}}

\code{\link{dropSproc}}

\code{\link{checkSproc}}
}
}
