% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeInSQL.R
\name{executeSQLQuery}
\alias{executeSQLQuery}
\title{Execute a script in SQL}
\usage{
executeSQLQuery(connectionString, sqlQuery)
}
\arguments{
\item{connectionString}{character string. The connectionString to the database}

\item{sqlQuery}{character string. The query to execute}
}
\value{
The data frame returned by the query to the database
}
\description{
Execute a script in SQL
}
\examples{
\dontrun{
connection <- connectionInfo(database="AirlineTestDB")
executeSQLQuery(connection, sqlQuery="SELECT top 1 * from airline5000")
}


}
