% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlPackage.R
\name{sql_install.packages}
\alias{sql_install.packages}
\title{sql_install.packages}
\usage{
sql_install.packages(connectionString, pkgs, skipMissing = FALSE, repos,
  verbose = getOption("verbose"), scope = "private", owner = "",
  scriptFile = NULL)
}
\arguments{
\item{connectionString}{ODBC connection string to Microsoft SQL Server database.}

\item{pkgs}{character vector of the names of packages whose current versions should be downloaded from the repositories. If repos = NULL, a character vector of file paths of .zip files containing binary builds of packages. (http:// and file:// URLs are also accepted and the files will be downloaded and installed from local copies).}

\item{skipMissing}{logical. If TRUE, skips missing dependent packages for which otherwise an error is generated.}

\item{repos}{character vector, the base URL(s) of the repositories to use.Can be NULL to install from local files, directories.}

\item{verbose}{logical. If TRUE, more detailed information is given during installation of packages.}

\item{scope}{character string. Should be either "public" or "private". "public" installs the packages on per database public location on SQL server which in turn can be used (referred) by multiple different users. "private" installs the packages on per database, per user private location on SQL server which is only accessible to the single user.}

\item{owner}{character string. Should be either empty '' or a valid SQL database user account name. Only 'dbo' or users in 'db_owner' role for a database can specify this value to install packages on behalf of other users. A user who is member of the 'db_owner' group can set owner='dbo' to install on the "public" folder.}

\item{scriptFile}{character string - a file where to record the tsql that is run by the function.}
}
\value{
invisible(NULL)
}
\description{
Installs R packages on a SQL Server database. Packages are downloaded on the client and then copied and installed to SQL Server into "public" and "private" folders. Packages in the "public" folders can be loaded by all database users running R script in SQL. Packages in the "private" folder can be loaded only by a single user. 'dbo' users always install into the "public" folder. Users who are members of the 'db_owner' role can install to both "public" and "private" folders. All other users can only install packages to their "private" folder.
}
\seealso{
{
\code{\link{sql_remove.packages}} to remove packages

\code{\link{sql_installed.packages}} to enumerate the installed packages

\code{\link{install.packages}} for the base version of this function
}
}
