% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlPackage.R
\name{sql_installed.packages}
\alias{sql_installed.packages}
\title{sql_installed.packages}
\usage{
sql_installed.packages(connectionString, priority = NULL,
  noCache = FALSE, fields = "Package", subarch = NULL,
  scope = "private", owner = "", scriptFile = NULL)
}
\arguments{
\item{connectionString}{ODBC connection string to Microsoft SQL Server database.}

\item{priority}{character vector or NULL (default). If non-null, used to select packages; "high" is equivalent to c("base", "recommended"). To select all packages without an assigned priority use priority = "NA".}

\item{noCache}{logical. If TRUE, do not use cached information, nor cache it.}

\item{fields}{a character vector giving the fields to extract from each package's DESCRIPTION file, or NULL. If NULL, the following fields are used: "Package", "LibPath", "Version", "Priority", "Depends", "Imports", "LinkingTo", "Suggests", "Enhances", "License", "License_is_FOSS", "License_restricts_use", "OS_type", "MD5sum", "NeedsCompilation", and "Built". Unavailable fields result in NA values.}

\item{subarch}{character string or NULL. If non-null and non-empty, used to select packages which are installed for that sub-architecture}

\item{scope}{character string which can be "private" or "public".}

\item{owner}{character string of a user whose private packages shall be listed (availableto dbo or db_owner users only)}

\item{scriptFile}{character string - a file where to record the tsql that is run by the function.}
}
\value{
matrix with enumerated packages
}
\description{
Enumerates the currently installed R packages on a SQL Server for the current database
}
\seealso{
{
\code{\link{sql_install.packages}} to install packages

\code{\link{sql_remove.packages}} to remove packages

\code{\link{installed.packages}} for the base version of this function

}
}
