% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlPackage.R
\name{sql_remove.packages}
\alias{sql_remove.packages}
\title{sql_remove.packages}
\usage{
sql_remove.packages(connectionString, pkgs, dependencies = TRUE,
  checkReferences = TRUE, verbose = getOption("verbose"),
  scope = "private", owner = "", scriptFile = NULL)
}
\arguments{
\item{connectionString}{ODBC connection string to SQL Server database.}

\item{pkgs}{character vector of names of the packages to be removed.}

\item{dependencies}{logical. If TRUE, does dependency resolution of the packages being removed and removes the dependent packages also if the dependent packages aren't referenced by other packages outside the dependency closure.}

\item{checkReferences}{logical. If TRUE, verifies there are no references to the dependent packages by other packages outside the dependency closure. Use FALSE to force removal of packages even when other packages depend on it.}

\item{verbose}{logical. If TRUE, more detailed information is given during removal of packages.}

\item{scope}{character string. Should be either "public" or "private". "public" removes the packages from a per-database public location on SQL Server which in turn could have been used (referred) by multiple different users. "private" removes the packages from a per-database, per-user private location on SQL Server which is only accessible to the single user.}

\item{owner}{character string. Should be either empty '' or a valid SQL database user account name. Only 'dbo' or users in 'db_owner' role for a database can specify this value to remove packages on behalf of other users. A user who is member of the 'db_owner' group can set owner='dbo' to remove packages from the "public" folder.}

\item{scriptFile}{character string - a file where to record the tsql that is run by the function.}
}
\value{
invisible(NULL)
}
\description{
Removes R packages from a SQL Server database.
}
\seealso{
{
\code{\link{sql_install.packages}} to install packages

\code{\link{sql_installed.packages}} to enumerate the installed packages

\code{\link{remove.packages}} for the base version of this function

}
}
